:recycle: Parse the config in application
